+++ internal/config_test.go
@@ -9,6 +9,68 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
+func TestConfig_tls(t *testing.T) {
+	t.Run("with no ENV", func(t *testing.T) {
+		var calls int
+		warn = func(msg string, args ...any) {
+			calls++
+		}
+		usingProgramArgs(t, "thruster", "echo", "hello")
+
+		c, err := NewConfig()
+		require.NoError(t, err)
+
+		assert.Equal(t, calls, 0)
+		assert.Equal(t, "", c.TLSDomain)
+	})
+
+	t.Run("with deprecated SSL_DOMAIN", func(t *testing.T) {
+		var warns []string
+		warn = func(msg string, args ...any) {
+			warns = append(warns, msg)
+		}
+		usingProgramArgs(t, "thruster", "echo", "hello")
+		usingEnvVar(t, "SSL_DOMAIN", "example.com")
+
+		c, err := NewConfig()
+		require.NoError(t, err)
+
+		assert.Equal(t, warns, []string{"SSL_DOMAIN is deprecated. Use TLS_DOMAIN instead."})
+		assert.Equal(t, "example.com", c.TLSDomain)
+	})
+
+	t.Run("with TLS_DOMAIN", func(t *testing.T) {
+		var calls int
+		warn = func(msg string, args ...any) {
+			calls++
+		}
+		usingProgramArgs(t, "thruster", "echo", "hello")
+		usingEnvVar(t, "TLS_DOMAIN", "example.com")
+
+		c, err := NewConfig()
+		require.NoError(t, err)
+
+		assert.Equal(t, calls, 0)
+		assert.Equal(t, "example.com", c.TLSDomain)
+	})
+
+	t.Run("with both SSL_DOMAIN and TLS_DOMAIN", func(t *testing.T) {
+		var warns []string
+		warn = func(msg string, args ...any) {
+			warns = append(warns, msg)
+		}
+		usingProgramArgs(t, "thruster", "echo", "hello")
+		usingEnvVar(t, "SSL_DOMAIN", "ssl.example.com")
+		usingEnvVar(t, "TLS_DOMAIN", "tls.example.com")
+
+		c, err := NewConfig()
+		require.NoError(t, err)
+
+		assert.Equal(t, warns, []string{"SSL_DOMAIN is deprecated. Use TLS_DOMAIN instead."})
+		assert.Equal(t, "tls.example.com", c.TLSDomain)
+	})
+}
+
 func TestConfig_defaults(t *testing.T) {
 	usingProgramArgs(t, "thruster", "echo", "hello")